Finalize Magick.NET migration and validations
Agent-Logs-Url: https://github.com/Pepsi1x1/DocxTemplater/sessions/bd8b59b8-a9e0-4593-ad5b-0ad0a3de5af9

+++ DocxTemplater.Images/ImageService.cs
@@ -75,7 +75,7 @@ namespace DocxTemplater.Images
                         {
                             throw new OpenXmlTemplateException("Could not find a valid image part");
                         }
-                        imageInfoInformation = new ImageInformation(image.Width, image.Height, imagePartRelId, exifRotation);
+                        imageInfoInformation = new ImageInformation((int)image.Width, (int)image.Height, imagePartRelId, exifRotation);
                         m_imagePartRelIdCache[imageBytes] = imageInfoInformation;
                     }
                 }
